Move utils to util package, info types to info package
