+++ src/main/java/me/darknet/resconstruct/info/MethodMember.java
@@ -1,4 +1,4 @@
-package me.darknet.resconstruct;
+package me.darknet.resconstruct.info;
 
 public class MethodMember extends Member{
 	public MethodMember(int access, String name, String desc) {
