+++ src/main/java/me/darknet/resconstruct/util/TypeUtils.java
@@ -1,8 +1,8 @@
-package me.darknet.resconstruct;
+package me.darknet.resconstruct.util;
 
 import org.objectweb.asm.Type;
 
-public class Poggers {
+public class TypeUtils {
     public static int getArgumentsSize(String desc) {
         int size = 0;
         Type methodType = Type.getMethodType(desc);
